- reset fang/eye ordering - AppearanceScreen only fro vampire - player model not scaling coords
